dirw: symlinks to files.
The direwalker is currently buggy because it trusts the dirent to
tell it whether to treat something as a file or directory.  This
is not guarnateed two work on all filesystems, and fails on
other cases like symlinks.

This change is mostly to the test-case generator, so that
you can actually put symlinks into your test case.

The current test-case has a single simlink to a file.  And
so the dirwalker passes the test by trivially treating all
files as symlinks.