doc: split make_stub_
This is the thing that actually converts the dirent into a stub_,
fixing up filenames, normalising filetypes and reporting errors
as needed.

There is a slight behaviour change in this cleanup: now we
call stat() for anything that is not a file or directory.
So if the dirent wrongly claims something is FIFO when it is
(or becomes) a regular file, we will trust stat().