clean: root-as-file works (too permissively).
make_stub_ now has become make_stub_from_de, to distinguish it from the new
function make_stub_from_path.  The latter is now pointed at whatever root is
supplied in order to kick of the recursion (which previously always was
read_dir).

This means that we ALWAYS accept directories and also files (the latter is
untested).  There used to be a sad-case test fails when the root is a file.
I have commented it out, but I want to re-store it when the root-as-file
feature is made into an option.